refactor(web): move environment-store to stores/ and error-report to utils/
environment-store.ts is a Zustand store with createSelectors() — it
belongs in stores/, not lib/. error-report.ts is a shared utility
function — it belongs in utils/, not lib/. Per CONVENTIONS.md, lib/
is reserved for third-party integrations and infrastructure.
