+++ apps/web/src/domains/settings/components/panels/assistant-terminal-panel.tsx
@@ -5,7 +5,7 @@ import { Dropdown } from "@vellum/design-library/components/dropdown";
 import { TerminalPanel } from "@/components/terminal-panel";
 import type { MaintenanceMode } from "@/generated/api/types.gen";
 import { getAssistant } from "@/assistant/api";
-import { reportError } from "@/lib/error-report";
+import { reportError } from "@/utils/error-report";
 
 type TerminalService = "assistant" | "gateway" | "credential-executor";
