+++ apps/web/src/domains/settings/components/panels/debug-controls-panel.tsx
@@ -9,7 +9,7 @@ import { RestartAssistant } from "@/domains/settings/components/restart-assistan
 import { RecoveryModeControls } from "@/domains/settings/components/recovery-mode-controls";
 import { type Assistant, getAssistant } from "@/assistant/api";
 import { useAuthStore } from "@/stores/auth-store";
-import { reportError } from "@/lib/error-report";
+import { reportError } from "@/utils/error-report";
 import { clearOnboardingFlags } from "@/lib/onboarding-cleanup";
 import { routes } from "@/utils/routes";
